Add output file support and benchmark filtering
